app/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
sqlalchemy>=1.4.0
psycopg2-binary
python-dotenv
pillow>=10.0.0
jinja2>=3.0.0
requests>=2.28.1